[FIX] floating state was sometimes wrong when windows being sent across workspaces.
This was due to a variable (hvar) being declared as an int, later in the program i tested if hvar was empty.
since the default value for int in bash is 0, that test alwasy passed. Fixed by changing type to normal var.

Fixes github issue #99

---
also fixed a template bug so that environment variables are now included.