[FIX] better handling of multiple criteria
This means that options must have arguments (--title TITLE ...).
To display the list, simply pass the word LIST as the argument
to the criterion option.

Example
--class LIST
will list all visible windows and display the class name in
the last field of the table.

Directions (u|l|d|r) can be used together with LIST
to effect the debug variables.