[GLOBAL] copy stderr to fd 3
Always print output of --help and --version to fd 3.
Reasoning for this is that if BASHBUD_LOG is set all stderr output will
get redirected to a log file. And I realized that --help and --version
output is not wanted in the log, and always wanted on the regular stderr (terminal).

Also removed un-necessary touch when log file was set