[FIX] add dummy commands to rule without a command
This can happen if a rule is incorrectly declared (#207) , where f.i the expected
linebreak that separates rule and command is escaped.

This commit increment command_counter for such lines, and adds a "nop" command as
a place holder. It will also print a message both when such rules are parsed and executed.

fixes #207